Move booking query to a separate service
