+++ SpotlessSolutions.Web/ClientApp/src/pages/dashboard/components/modals/BookingsDetailModal.style.scss
@@ -53,7 +53,7 @@
           overflow-wrap: anywhere;
 
           & select {
-            @apply -ml-1;
+            @apply -ml-1 p-1;
           }
         }
