+++ SpotlessSolutions.Web/Services/Bookings/BookingManager.cs
@@ -60,116 +60,6 @@ public class BookingManager : IBookingManager
         return true;
     }
 
-    public async Task<IEnumerable<BookingObject>> GetBooking(int year, int month)
-    {
-        var start = new DateTime(year, month, 1).ToUniversalTime();
-        var end = new DateTime(year, month, DateTime.DaysInMonth(year, month))
-            .ToUniversalTime();
-
-        var result = (await GetBooking(start, end)).ToList();
-        return result;
-    }
-
-    public async Task<IEnumerable<BookingObject>> GetBooking(DateTime start, DateTime end)
-    {
-        var bookings = await _context.Bookings
-            .Include(x => x.User.User)
-            .Include(x => x.Address)
-            .Where(x => x.Schedule >= start && x.Schedule <= end)
-            .ToArrayAsync();
-
-        return bookings
-            .Where(x =>
-            {
-                var service = _registry.GetActivatedServiceInstance(x.MainServiceId);
-                if (service == null)
-                {
-                    return false;
-                }
-
-                return x.MainServiceConfiguration.GetCalculationParams() != null;
-            })
-            .Select(x =>
-            {
-                var mainService = _registry.GetActivatedServiceInstance(x.MainServiceId)!;
-                var mainServiceParams = x.MainServiceConfiguration.GetCalculationParams()!;
-                var mainServiceCalculation = mainService.Calculate(mainServiceParams);
-                
-                var mainServiceDetail = new ServiceDetailConfig
-                {
-                    Service = new ServiceDetails
-                    {
-                        Name = mainService.GetName(),
-                        Description = mainService.GetDescription(),
-                        Id = mainService.GetId(),
-                        Type = mainService.GetServiceType()
-                    },
-                    BookingDescriptor = mainServiceCalculation.Descriptors,
-                    Calculated = mainServiceCalculation.CalculatedValue
-                };
-
-                var addons = MapAddons(x);
-
-                var user = new User
-                {
-                    FirstName = x.User.FirstName,
-                    LastName = x.User.LastName,
-                    UserId = x.User.Id
-                };
-
-                var address = new Address
-                {
-                    Barangay = x.Address.Barangay,
-                    City = x.Address.City,
-                    District = x.Address.District,
-                    PostalCode = x.Address.PostalCode,
-                    Province = x.Address.Province,
-                    Street = x.Address.Street
-                };
-
-                var status = x.Status;
-                
-                return new BookingObject(x.Id, x.Schedule, mainServiceDetail, addons, status, user, address, x.FinalPrice);
-            });
-    }
-
-    private List<ServiceDetailConfig> MapAddons(Booking booking)
-    {
-        return booking.Addons
-            .Where(data =>
-            {
-                var instance = _registry.GetActivatedServiceInstance(data.Key);
-                if (instance == null)
-                {
-                    return false;
-                }
-
-                return instance.GetServiceType() == ServiceType.Addons &&
-                       data.Value.GetCalculationParams() != null;
-            })
-            .Select(data =>
-            {
-                var service = _registry.GetActivatedServiceInstance(data.Key)!;
-                var config = data.Value.GetCalculationParams()!;
-
-                var calculation = service.Calculate(config);
-
-                return new ServiceDetailConfig
-                {
-                    Service = new ServiceDetails
-                    {
-                        Name = service.GetName(),
-                        Description = service.GetDescription(),
-                        Id = service.GetId(),
-                        Type = service.GetServiceType(),
-                    },
-                    BookingDescriptor = calculation.Descriptors,
-                    Calculated = calculation.CalculatedValue
-                };
-            })
-            .ToList();
-    }
-
     public Task<IEnumerable<DateTime>> GetAnonymizedBookingDetails(int year, int month)
     {
         var start = new DateTime(year, month, 1).ToUniversalTime();
